requirements.txt
Django
django-oauth-toolkit
django-cors-headers
Pillow
sentry-sdk[django]
django-hitcount